requirements.txt
# Requirements
# pip install -r requirements.txt
#

django==2.2.11
django-extensions
paramiko
djangorestframework
validators
netaddr
